Add Zoom To Plugin
Added a new Zoom To plugin. This plugin is dynamic action that pans and zooms to the extent of a GeoJSON format geometry from a sql query.

Raster Plugin: Removed spinner since it doesn't work consistently. 

Assigned non-Subversion version numbers to all plugins and moved subversion commit information and changes to the comments. These files were committed to internal Subversion.