Add files via upload
Added full Geolocation support to Drawing Controls. In addition to existing functionality, if a user clicks the geolocation button when in line or polygon edit mode, a vertex will be added at the users location. The geolocation button is visible when any draw tool is active or when a vertex is selected.
Moved the Drawing Controls javascript code related to styles to a separate file.