Maplibre and APEX 22 accommodation
Checked all plugins against application express version 22.1. In this version, Oracle swapped mapbox for maplibre. This required changes in the javascript to use the correct library and, when a control is added, to change the classname of the control group element.

Modified REST plugin to ignore isSourceLoaded. That wasn't working.