WMS fix for APEX 22.2
Fixed WMS Layer to work on APEX 22.2. It appears adding the 'TE' header value of 'trailer' to HTTP calls, which was in previous working versions but not in 22.2, fixes a CORS or CORS-like issue. 
For WMS Layer, use the 'idle' event as a fallback to remove the data loading spinner.