New legend plugin and event handling refactor.
Added Region Plugin for a Legend. Added a page in the demo application for the Legend.. Bumped version of application and plugins to 4.5. Refactored use of event handlers in the plugins. The Set Custom Marker and Zoom To plugins use an interval to "wait" a valid map before executing their intended functions. The item plugins such as WMS, GeoJSON, Drawing, etc. no longer use 'render' or 'load' callbacks for initialization tasks since they run inside of a 'spatialmapinitialized' callback, making them redundant. Fixed Stationing plugin through event handler refactoring, restoring the display of the stationing layer and formatting of the chart station labels.