Mapbits Draw: First Render
Fixed problems in Mapbits Draw. The geometry was not rendering in read-only mode. Fixed by setting the default geometry mode to 'NONE' instead of empty string. Also, a error would be raised in APEX 23 on the first render of the drawing geometry. Fixed by trapping errors from map.jumpTo (and printing a console message).